pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-folder-organizer"
version = "0.1.0"
description = "LLM-based folder organizer (DeepSeek/OpenAI compatible)"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "openai>=1.0.0",
  "tqdm>=4.0.0",
  "streamlit>=1.30.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
lfo = "llm_folder_organizer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
build = [
  "pyinstaller>=6.0.0",
]

